move activity, fragment to package and fix comment
