+++ app/src/at-dinhtruong/java/asiantech/internship/summer/activityandfragment/LoginFragment.java
@@ -1,4 +1,4 @@
-package asiantech.internship.summer;
+package asiantech.internship.summer.activityandfragment;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -12,6 +12,7 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 
+import asiantech.internship.summer.R;
 import asiantech.internship.summer.utils.Validate;
 
 public class LoginFragment extends Fragment implements View.OnClickListener {
@@ -51,6 +52,7 @@ public class LoginFragment extends Fragment implements View.OnClickListener {
                     ((LoginActivity) getActivity()).setTitleToolBar(getString(R.string.signup));
                     ((LoginActivity) getActivity()).setButtonBack(R.drawable.ic_arrow_back_black_36dp);
                 }
+                assert getFragmentManager() != null;
                 FragmentTransaction fragmentTransaction = getFragmentManager().beginTransaction();
                 fragmentTransaction.setCustomAnimations(R.anim.right_to_left1, R.anim.right_to_left2, R.anim.left_to_right1, R.anim.left_to_right2);
                 fragmentTransaction.add(R.id.fragment_container, new SignUpFragment());
