+++ app/src/at-dinhtruong/java/asiantech/internship/summer/utils/Validate.java
@@ -4,8 +4,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Validate {
-    private final String PASSWORD_PATTERN = "^[a-zA-Z0-9]{7,}$";
-    private final String EMAIL_PATTERN = "^([a-zA-Z0-9_\\-.]+)@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.)|(([a-zA-Z0-9\\-]+\\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(]?)$";
+    private static final String PASSWORD_PATTERN = "^[a-zA-Z0-9]{7,}$";
+    private static final String EMAIL_PATTERN = "^([a-zA-Z0-9_\\-.]+)@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.)|(([a-zA-Z0-9\\-]+\\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(]?)$";
 
     public boolean isValidPassword(String passWord) {
         Pattern patternPassword;
