+++ app/src/at-dinhtruong/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".LoginActivity">
+    tools:context=".activityandfragment.LoginActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolBar"
