+++ app/src/at-dinhtruong/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:gravity="center"
     android:background="@color/colorVioletBlue"
     android:orientation="vertical"
-    tools:context=".LoginFragment">
+    tools:context=".activityandfragment.LoginFragment">
 
     <android.support.v7.widget.AppCompatEditText
         android:id="@+id/edtEmail"
